tambah sistem notifikasi sma ad perbaiki bbrp brg kecil

+++ database/migrations/2025_07_05_050011_create_jenis_pemeriksaans_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->boolean('pemakaianKontras');
             $table->integer('lamaPemeriksaan');
             $table->boolean('diDampingiDokter');
-            $table->double('harga');
+            $table->integer('harga');
             $table->timestamps();
         });
     }
